pyproject.toml
[tool.poetry]
name = "project-viewets-and-generics"
version = "0.1.0"
description = ""
authors = ["nikolyai-tarasov <kolya.tarasov333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
sqlparse = "^0.5.3"
ipython = "^8.30.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
